refactor: 지도 아이콘 zoom 스케일링 적용 및 TooltipBox 유틸 분리
- 확대 시 아이콘(원, X, 다이아몬드, S/E 마커)이 화면상 고정 크기 유지하도록 / zoom 처리
- TelemetryMapView, LogMapCanvas 모두 적용
- 중복된 TooltipBox, TooltipData, TooltipInfo, formatElapsed를 MapTooltip.tsx로 분리
